fix all register controller bugs

+++ src/utils/cloudinary.js
@@ -1,6 +1,6 @@
 import { v2 as cloudinary } from 'cloudinary';
 import fs from "fs"
-import { v2 as cloudinary } from 'cloudinary';
+ 
 
  
     // Configuration
@@ -15,16 +15,21 @@ const uploadOnCloudinary=async(localFilePath)=> {
     try{
         if(!localFilePath) return null
         // upload on cloudinary
-        const response=await cloudinary.v2.uploader.upload(localFilePath, {
+        const response=await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
         })
         // file has been uploaded
-        console.log("file is uploaded on cloudinary", response.url)
+        fs.unlinkSync(localFilePath)
 
         return response
     } catch(error){
-        fs.unlinkSync(localFilePath) // remove the locally saved temp file as the upload operation got failed
-        return null
+        console.error("Cloudinary Error:", error);
+
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
+        return null;
     }
 }
